requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
idna==3.3
importlib-metadata==4.11.2
iniconfig==1.1.1
more-itertools==8.12.0
msedge-selenium-tools==3.141.4
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
requests==2.27.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.8
wcwidth==0.2.5
webdriver-manager==3.5.3
xlrd==2.0.1
zipp==3.7.0
names==0.3.0